implemented correct double vanishing point

+++ ScaleCalculator/ScaleCalculator/Measures.cpp
@@ -1,7 +1,7 @@
 #include "Measures.h"
 #include <cmath>
 
-Point::Point(int x, int y) {
+Point::Point(double x, double y) {
 	this->x = x;
 	this->y = y;
 }
